+++ orm-patterns-instances/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - The delete method for the phones to hibernate-jpa examples project.
 - The find service execution method to hibernate-jpa examples project.
 - The update service execution method to hibernate-jpa examples project.
+- The delete service execution method to hibernate-jpa examples project.
 
 ## 2020-11-09